Внес изменения: Rebase на ветку feature/M3L8-31-MardownToTextViewParser, добавил generic вместо протокола IVisitable, пофиксил PdfPrinterItem
