+++ MdEditor/Sources/Presentation/Flows/BrowseHTMLFlow/BrowseHTMLCoordinator.swift
@@ -12,9 +12,6 @@ import UIKit
 protocol IBrowseHTMLCoordinator: ICoordinator {
 	/// Уведомляет о закрытии сцены просмотра файла.
 	func didCloseBrowseHTMLScene()
-	/// Открывает документ в формате pdf в новом окне.
-	/// - Parameter file: Принимает файл для откыртия.
-	func openPdf(_ file: DirectoryObject)
 }
 
 /// Координатор флоу просмотра файла в виде HTML-страницы..
@@ -52,18 +49,6 @@ final class BrowseHTMLCoordinator: IBrowseHTMLCoordinator {
 	func start() {
 		let browseHTMLViewController = BrowseHTMLAssembly.assemble(coordinator: self, file: file)
 		navigationController.show(browseHTMLViewController, sender: nil)
-	}
-
-	/// Открывает документ в формате pdf в новом окне.
-	/// - Parameter file: Принимает файл для откыртия.
-	func openPdf(_ file: DirectoryObject) {
-		let browsePdfCoordinator = BrowsePDFCoordinator(
-			navigationController: navigationController,
-			finishDelegate: self,
-			file: file
-		)
-		childCoordinators.append(browsePdfCoordinator)
-		browsePdfCoordinator.start()
 	}
 }
